mobx: Move NavItem out of NavStore and remove lodash dependancy
